Refactor package
* Moved parts to pipliceses pkg
* Fixed tests
* Fixed workflow
* Updated gitignore
* Updated setup.cfg
