+++ .github/workflows/python-package.yml
@@ -24,9 +24,10 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         if [ -f dev-requirements.txt ]; then pip install -r dev-requirements.txt; fi
+        pip install -e .
     - name: Test with pytest
       run: |
-        pytest --pycodestyle -v --cov --cov-report=xml
+        pytest --pycodestyle -v --cov piplicenses --cov-report=xml
     - name: Upload code coverage to Codecov
       if: ${{ matrix.python-version==3.8 }}
       uses: codecov/codecov-action@v1
