+++ piplicenses/const.py
@@ -1,0 +1,76 @@
+try:
+    from prettytable.prettytable import ALL
+    PTABLE = True
+except ImportError:  # pragma: no cover
+    PTABLE = False
+
+
+__pkgname__ = 'pip-licenses'
+__version__ = '3.2.0'
+__author__ = 'raimon'
+__license__ = 'MIT'
+__summary__ = (
+    'Dump the software license list of '
+    'Python packages installed with pip.'
+)
+__url__ = 'https://github.com/raimon49/pip-licenses'
+
+
+FIELD_NAMES = (
+    'Name',
+    'Version',
+    'License',
+    'LicenseFile',
+    'LicenseText',
+    'NoticeFile',
+    'NoticeText',
+    'Author',
+    'Description',
+    'URL',
+)
+
+
+SUMMARY_FIELD_NAMES = (
+    'Count',
+    'License',
+)
+
+
+DEFAULT_OUTPUT_FIELDS = (
+    'Name',
+    'Version',
+)
+
+
+SUMMARY_OUTPUT_FIELDS = (
+    'Count',
+    'License',
+)
+
+
+METADATA_KEYS = (
+    'home-page',
+    'author',
+    'license',
+    'summary',
+    'license_classifier',
+)
+
+# Mapping of FIELD_NAMES to METADATA_KEYS where they differ by more than case
+FIELDS_TO_METADATA_KEYS = {
+    'URL': 'home-page',
+    'Description': 'summary',
+    'License-Metadata': 'license',
+    'License-Classifier': 'license_classifier',
+}
+
+
+SYSTEM_PACKAGES = (
+    __pkgname__,
+    'pip',
+    'PTable' if PTABLE else 'prettytable',
+    'setuptools',
+    'wheel',
+)
+
+LICENSE_UNKNOWN = 'UNKNOWN'
