+++ setup.cfg
@@ -48,4 +48,15 @@ universal = 0
 test = pytest
 
 [tool:pytest]
-addopts = --pycodestyle -v --cov --cov-report term-missing
+testpaths = tests
+addopts = --pycodestyle -v --cov piplicenses --cov-report term-missing
+
+[pycodestyle]
+max-line-length = 119
+
+[tool:isort]
+# https://github.com/timothycrosley/isort
+multi_line_output = 4
+line_length = 72
+known_first_party =
+    piplicenses
